[Silver III] Title: 주유소, Time: 100 ms, Memory: 6132 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -5,7 +5,7 @@ int n;
 vector<long long> load;
 vector<long long> city;
 vector<long long> price;
-int ans;
+long long ans;
 
 int main(void) {
     cin >> n;
@@ -35,5 +35,4 @@ int main(void) {
         ans += price[i];
     }
     cout << ans;
-    // 마지막 기름
 }
